Don't trust session's user_id when logging in
If performing a login, do not short circuit the CAS machinery when
the session already contains an user id. The session may be stale,
and this way we are losing the back_url.

In a nutshell, if a login is requested, do the login, don't try to
be too smart and avoid it.